pyproject.toml
[project]
name = "betopia-rag-mvp"
version = "0.1.0"
description = "MVP RAG system for Betopia.ai - returns EvidencePackage for outer agent."
requires-python = ">=3.10"
dependencies = [
    "qdrant-client>=1.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
]

[project.optional-dependencies]
local-embeddings = [
    "sentence-transformers>=2.7.0",
]
postgres = [
    "psycopg[binary,pool]>=3.1.0",
]
reranker = [
    "sentence-transformers>=2.7.0",
]
ui = [
    "streamlit>=1.36.0",
]
markitdown = [
    # Microsoft markitdown: PDF/DOCX/XLSX/PPTX/HTML/CSV -> clean markdown
    # with headings preserved. Without this extra, pypdf is used for PDFs
    # and only .txt/.md/.pdf are supported.
    "markitdown[all]>=0.1.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
rag-ingest = "rag.cli.ingest:main"
rag-query = "rag.cli.query:main"
rag-api = "rag.api.server:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["rag*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# FastAPI dependency-injection pattern: File(...), Form(...), Depends(...) used as defaults.
"rag/api/app.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
